+++ src/Glpi/Controller/Config/Helpdesk/UpdateTileController.php
@@ -107,7 +107,7 @@ final class UpdateTileController extends AbstractController
         $tiles = $this->tiles_manager->getTiles(new SessionInfo(
             profile_id: $profile_id,
         ), check_availability: false);
-        return $this->render('pages/admin/helpdesk_home_config_tiles.html.twig', [
+        return $this->render('pages/admin/profile/helpdesk_home/tiles.html.twig', [
             'tiles_manager' => $this->tiles_manager,
             'tiles' => $tiles,
         ]);